JSP+SERVLET

+++ src/soac/miniprojet/servlets/StudentsServlet.java
@@ -42,6 +42,7 @@ public class StudentsServlet extends HttpServlet {
 
             request.setAttribute("user",user.getNom()+" "+user.getPrenom());
             request.setAttribute("role",user.getRole());
+            System.out.println(user.getRole());
 
             LinkedList<Students> students = new StudentsApi().getAll();
             request.setAttribute("students", students);
@@ -59,34 +60,58 @@ public class StudentsServlet extends HttpServlet {
      * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
      */
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String _id = request.getParameter("id");
-        int id = Integer.parseInt(_id);
-        String nom = request.getParameter("nom");
-        String  prenom = request.getParameter("prenom");
-        String  date_naiss = request.getParameter("date_naiss");
-        String  sexe = request.getParameter("sexe");
-        String  num_bac = request.getParameter("num_bac");
+        HttpSession session = request.getSession();
+        if (session.getAttribute("user") != null) {
+            Employees user = (Employees)session.getAttribute("user");
 
-        Students s = new Students();
-        s.setId(id);
-        s.setNom(nom);
-        s.setPrenom(prenom);
-        s.setDateNaiss(new Date(date_naiss));
-        s.setSexe(sexe);
-        s.setNumBac(num_bac);
-        s.setDateInsc(new Date());
-        api.update(s);
+            request.setAttribute("user",user.getNom()+" "+user.getPrenom());
+            request.setAttribute("role",user.getRole());
 
-        doGet(request, response);
+            String _id = request.getParameter("id");
+            int id = Integer.parseInt(_id);
+            String nom = request.getParameter("nom");
+            String  prenom = request.getParameter("prenom");
+            String  date_naiss = request.getParameter("date_naiss");
+            String  sexe = request.getParameter("sexe");
+            String  num_bac = request.getParameter("num_bac");
+
+            Students s = new Students();
+            s.setId(id);
+            s.setNom(nom);
+            s.setPrenom(prenom);
+            s.setDateNaiss(new Date(date_naiss));
+            s.setSexe(sexe);
+            s.setNumBac(num_bac);
+            s.setDateInsc(new Date());
+            api.update(s);
+
+            doGet(request, response);
+
+        } else response.sendRedirect(request.getContextPath() + "/login");
+
+
+
+
     }
 
 
 
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        super.doDelete(req, resp);
-       String id = req.getParameter("id");
-       api.deleteById(Integer.parseInt(id));
-       doGet(req,resp);
+        HttpSession session = req.getSession();
+        if (session.getAttribute("user") != null) {
+            Employees user = (Employees)session.getAttribute("user");
+
+            req.setAttribute("user",user.getNom()+" "+user.getPrenom());
+            req.setAttribute("role",user.getRole());
+
+            String id = req.getParameter("id");
+            api.deleteById(Integer.parseInt(id));
+            doGet(req,resp);
+
+        } else resp.sendRedirect(req.getContextPath() + "/login");
+
+
+
     }
 }
